Update Calls and Notes History
In the file:
custom/modules/ProspectLists/callModeView.php :

renderLogCall function is called on each Contact record.  This function is in
modules/ATC_Teleseller/ATC_CallLogger.php

This captures the call and note history for the current Contact.

This is encoded in JSON and saved on the page, as it loads.

When you click 'log call' or 'create notes' on the calling mode view, these values are passed in, and rendered onto a dialog by the javascript functions: LogCall or CreateNote

these javascript functions are located in modules/ATC_Teleseller/javascript/ATC_JS/call_logging.js

I would like to modify this so that the history is loaded when 'Log Call' or 'Create Note' are called, instead of ahead of time.  in the first line of the CreateNote function, for example, parses the note history json passed in.  So replace this with real-time loading of the data, instead of pre-loading.

Please use the test branch as starting point for this, as I've already done some work modifying the callModeView.php view to use SQL instead of the bean, to speed up loading times.

We will be looking to replace this in the future, so please don't bother spending time updating other items.